persistence of modal & opacity fix
