+++ src/App.js
@@ -18,7 +18,26 @@ function App() {
   const [messageCount, setMessageCount] = useState({});
   const [modalPerson, setModalPerson] = useState(null);
 
+  const currentUser = JSON.parse(localStorage.getItem('user'));
+  //to set Object for blurry images dpending on messageCount
+  useEffect(() => {
+    const fetchMessageCounts = async () => {
+        try {
+            const messageCounts = {};
+            for (const match of matches) {
+                const response = await axios.get(`http://localhost:8080/messages/${currentUser.stringId}/${match.stringId}`, { withCredentials: true });
+                messageCounts[match.stringId] = response.data.length;
+            }
+            setMessageCount(messageCounts);
+        } catch (error) {
+            console.error("Failed to fetch message counts", error);
+        }
+    };
 
+    if (currentUser && matches) {
+        fetchMessageCounts();
+    }
+}, [currentUser, matches]);
 
   const handleLogout = () => {
     setUser(null);
@@ -38,7 +57,7 @@ function App() {
       window.location.pathname === '/matches' ||
       window.location.pathname === '/messages' ||
       window.location.pathname === '/profile') && (
-        <Nav handleLogout={handleLogout} />
+        <Nav handleLogout={handleLogout} setModal={setModal} />
       )
   }
         <Routes>
