+++ src/components/Modal/Modal.scss
@@ -1,5 +1,6 @@
 .Modal{
-    background-color: white;
-    font-weight: 600;
-    color: black;
+    // display: flex;
+    // justify-content: center;
+    // align-items: center;
+    z-index: 1000;
 }
